dl and ulspeed are now curl_off_t so typecast them to double when we return their values

+++ lib/getinfo.c
@@ -18,7 +18,7 @@
  * This software is distributed on an "AS IS" basis, WITHOUT WARRANTY OF ANY
  * KIND, either express or implied.
  *
- * $Id: getinfo.c,v 1.35 2004-03-10 16:20:33 bagder Exp $
+ * $Id: getinfo.c,v 1.36 2004-03-11 21:51:55 bagder Exp $
  ***************************************************************************/
 
 #include "setup.h"
@@ -141,10 +141,10 @@ CURLcode Curl_getinfo(struct SessionHandle *data, CURLINFO info, ...)
     *param_doublep = (double)data->progress.downloaded;
     break;
   case CURLINFO_SPEED_DOWNLOAD:
-    *param_doublep =  data->progress.dlspeed;
+    *param_doublep =  (double)data->progress.dlspeed;
     break;
   case CURLINFO_SPEED_UPLOAD:
-    *param_doublep = data->progress.ulspeed;
+    *param_doublep = (double)data->progress.ulspeed;
     break;
   case CURLINFO_SSL_VERIFYRESULT:
     *param_longp = data->set.ssl.certverifyresult;
